+++ lab1/src/vis.c
@@ -6,7 +6,7 @@
 #define MAXLENBUFFER 1000
 
 int main(int argc, char ** argv){
-  double realAverage = 0.0, imgAverage = 0.0, power = 0.0, noise = 0.0, k = 0.0;
+  float realAverage = 0.0, imgAverage = 0.0, power = 0.0, noise = 0.0, k = 0.0;
   char *buffer = (char *) calloc(MAXLENBUFFER, sizeof(char));
   float u, v, r, i, n;
   while(fscanf(stdin, "%f,%f,%f,%f,%f\n", &u, &v, &r, &i, &n) != EOF){
@@ -19,7 +19,7 @@ int main(int argc, char ** argv){
   }
   realAverage = realAverage/k;
   imgAverage = imgAverage/k;
-  sprintf(buffer, "%lf,%lf,%lf,%lf,%d", realAverage, imgAverage, power, noise, (int) k);
+  sprintf(buffer, "%f,%f,%f,%f,%d", realAverage, imgAverage, power, noise, (int) k);
   printf("%s",buffer);
   free(buffer);
   return 0;
